+++ include/roq/download_end.hpp
@@ -20,7 +20,7 @@ namespace roq {
 //! Identifies the end of a download sequence
 struct ROQ_PUBLIC DownloadEnd final {
   std::string_view account;    //!< Account name
-  uint32_t max_order_id = {};  //!< Highest previous order identifier (as seen by gateway)
+  uint64_t max_order_id = {};  //!< Highest previous order identifier (as seen by gateway)
 };
 
 template <>
